edited by SS
Made some changes:
1. added a loop for all substances while running optimise minimise. 
2. the minimise value in optimise_fn return statement - i had to change the df[waste] -> data[sheet_name]['Waste'][substance].astype(float
astype because recent pd.dataframe uses the datatype object and there is no sqrt of an object. so it needs to be float.
3. as an optional argument to optimise, instead of (c_ddr, df and f_avg), now we send 4 arguments c_ddr, data[sheet_name], f_avg, substance

Remark: In the excel file, magnesium is <0.21 which cannot be converted. you can put that as a "caution" in the excel file - include it in instructions to users "be careful to not have non-numeric values in the excel"
What i did was in data.xlsx, I changed that to 0.2
What a user could alternatively do if they have datapoint like '<0.21' is remove it.

To be done:
1. save the substance, k_Th and q_e for each sheet as a dataframe "expt". Save as csv for each experiment.
2. once the substance loop in line 98 finishes and you have this "expt" dataframe - make plots of 5 rows (solutes) and 2 columns (waste and downstream sorbents)